+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Anthropic: Show warning when generation config incompatible with extended thinking is used (affects `temperature`, `top_p`, and `top_k`).
 - AzureAI: Don't include `tools` or `tool_choice` in  requests when emulating tool calling (avoiding a 400 error).
 - Eval retry: Improve error message when attempting to retry tasks in packages that have not been registered.
-
+- Warn when a passed `--sample-id` is not found in the target dataset.
 
 ## v0.3.96 (13 May 2025)